Price modifier action menu styles.
